Renamed slice file
