+++ src/Redux/Store/Store.js
@@ -1,13 +1,13 @@
 import createSagaMiddleware from '@redux-saga/core';
 import { configureStore} from '@reduxjs/toolkit'
-import counterSlice from '../Reducers/counterSlice'; 
+import passwordSlice from '../Reducers/passwordSlice'; 
 import saga from '../Saga/firebaseSaga';
 
 const sagaMiddleware = createSagaMiddleware();
 const middleware = [sagaMiddleware];
 
 export const store = configureStore({
-  reducer: {counterSlice},
+  reducer: {passwordSlice},
   middleware
 });
