+++ src/components/AddButton/AddItemButton.js
@@ -3,7 +3,7 @@ import IconButton from '@mui/material/IconButton';
 import AddIcon from '@mui/icons-material/Add';
 import './AddItemButton.css';
 
-function AddItemButton({onClickHandler}) {
+const AddItemButton = ({onClickHandler}) =>{
     return (
         <div className='button-container'>
         <IconButton aria-label="add button" className='add-button' onClick={onClickHandler}>
@@ -13,4 +13,4 @@ function AddItemButton({onClickHandler}) {
     )
 }
 
-export default AddItemButton
+export default React.memo(AddItemButton)
